condense message format lib
